+++ src/features/home/components/module/diamon/style.scss
@@ -1,0 +1,57 @@
+[class^='item'] {
+    width: 200px;
+    height: 200px;
+    background: white;
+    transform: rotate(45deg);
+    margin: -20px 60px;
+    position: relative;
+    border-radius: 20px;
+    
+    &.hidden {
+        visibility: hidden;
+    }
+}
+
+.content {
+    transform: rotate(-45deg);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+    margin: auto;
+    left: 0;
+    position: absolute;
+    right: 0;
+    top: 0;
+    bottom: 0; 
+
+    &:hover {
+        img {
+            width: 90px;
+            margin-top: 60px;
+        }
+
+        span {
+            opacity: 0;
+        }
+    }
+
+    img {
+        width: 60px;
+        margin: -20px auto 5px;
+        transition: all 0.5s;
+    }
+
+    .title {
+        left: 0;
+        top: 0;
+        right: 0;
+        margin: 18px auto 0;
+        color: black;
+        text-shadow: unset;
+        font-size: 18px;
+        text-align: center;
+        font-weight: bold;
+        transition: all 0.5s;
+    }
+}
