+++ src/features/home/components/module/solution/style.scss
@@ -1,0 +1,34 @@
+.diamond-group {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    margin: 0;
+    padding: 0;
+    display: block;
+
+    .group {
+        display: flex;
+        justify-content: center;
+
+        &--three{
+            justify-content: space-around;
+        }
+        
+        &--four{
+            justify-content: flex-start;
+        }
+    }
+
+    .title {
+        text-align: center;
+
+        h1 {
+            font-size: 3.5rem;
+            font-weight: 700;
+            text-shadow: 1px 1px 10px;
+            width: 320px;
+            margin: auto;
+        }
+    }
+}
