+++ src/features/home/components/module/style.scss
@@ -1,0 +1,23 @@
+.module {
+    position: relative;
+
+    .main-module {
+        width: 100%;
+    }
+
+    .container {
+        position: absolute;
+        top: 10%;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        width: 60%;
+        color: white;
+        box-shadow: unset;
+    }
+
+    .giaiphap {
+        position: relative;
+        top: 15%;
+    }
+}
