+++ src/features/home/index.jsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { Row, Col } from 'react-bootstrap';
 import Slider from './components/slider/index';
 import LatestNews from './components/latestNew/index';
-import Module from './components/module/module'
+import Module from './components/module'
 import './style.scss';
 const Home = () => {
     const { t } = useTranslation();
